requirements.txt
# LOTR Data Cloud POC Dependencies

# Web Framework
Flask==3.0.0
Werkzeug==3.0.1

# HTTP Requests
requests==2.31.0

# Environment Management
python-dotenv==1.0.0

# Date/Time utilities
python-dateutil==2.8.2

